+++ src/main/java/ru/bunkov/calculation/api/external/accounting/dto/SearchAccountingDto.java
@@ -3,20 +3,22 @@ package ru.bunkov.calculation.api.external.accounting.dto;
 import lombok.Data;
 import ru.bunkov.calculation.model.accounting.Form;
 
+import java.math.BigDecimal;
+
 @Data
 public class SearchAccountingDto {
 
     private Form organizationalAndLegalForm;
 
-    private Double generalTaxationSystemMin;
+    private BigDecimal generalTaxationSystemMin;
 
-    private Double generalTaxationSystemMax;
+    private BigDecimal generalTaxationSystemMax;
 
-    private Double simplifiedTaxationSystemMin;
+    private BigDecimal simplifiedTaxationSystemMin;
 
-    private Double simplifiedTaxationSystemMax;
+    private BigDecimal simplifiedTaxationSystemMax;
 
-    private Double patentMin;
+    private BigDecimal patentMin;
 
-    private Double patentMax;
+    private BigDecimal patentMax;
 }
