+++ src/main/java/ru/bunkov/calculation/api/external/capitalbuilding/dto/CapitalBuildingObjDto.java
@@ -2,6 +2,7 @@ package ru.bunkov.calculation.api.external.capitalbuilding.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -10,5 +11,5 @@ public class CapitalBuildingObjDto {
 
     private String type;
 
-    private Double cost;
+    private BigDecimal cost;
 }
