+++ src/main/java/ru/bunkov/calculation/api/external/capitalbuilding/dto/CreateCapitalBuildingObjDto.java
@@ -2,10 +2,12 @@ package ru.bunkov.calculation.api.external.capitalbuilding.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CreateCapitalBuildingObjDto {
 
     private String type;
 
-    private Double cost;
+    private BigDecimal cost;
 }
