+++ src/main/java/ru/bunkov/calculation/api/external/capitalbuilding/dto/SearchCapitalBuildingObjDto.java
@@ -2,9 +2,11 @@ package ru.bunkov.calculation.api.external.capitalbuilding.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class SearchCapitalBuildingObjDto {
     private String type;
 
-    private Double cost;
+    private BigDecimal cost;
 }
