+++ src/main/java/ru/bunkov/calculation/api/external/equipment/dto/CreateEquipmentDto.java
@@ -2,10 +2,12 @@ package ru.bunkov.calculation.api.external.equipment.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CreateEquipmentDto {
 
     private String equipmentType;
 
-    private Double averageCost;
+    private BigDecimal averageCost;
 }
