+++ src/main/java/ru/bunkov/calculation/api/external/equipment/dto/EquipmentDto.java
@@ -2,6 +2,7 @@ package ru.bunkov.calculation.api.external.equipment.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -11,5 +12,5 @@ public class EquipmentDto {
 
     private String equipmentType;
 
-    private Double averageCost;
+    private BigDecimal averageCost;
 }
