+++ src/main/java/ru/bunkov/calculation/api/external/income/dto/SearchPotentialIncomeDto.java
@@ -2,13 +2,15 @@ package ru.bunkov.calculation.api.external.income.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class SearchPotentialIncomeDto {
     private String typeOfBusiness;
 
-    private Double potentialAnnualIncome;
+    private BigDecimal potentialAnnualIncome;
 
-    private Double rate;
+    private BigDecimal rate;
 
-    private Double totalPotentCost;
+    private BigDecimal totalPotentCost;
 }
