+++ src/main/java/ru/bunkov/calculation/model/calculation/capitalinvestment/TotalBuildingCost.java
@@ -3,6 +3,7 @@ package ru.bunkov.calculation.model.calculation.capitalinvestment;
 import lombok.*;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.util.List;
 
 @Embeddable
@@ -20,5 +21,5 @@ public class TotalBuildingCost {
     })
     private List<PairForBuilding> buildingCostIdAndAreaPairList;
 
-    private Double totalBuildingCost;
+    private BigDecimal totalBuildingCost;
 }
