+++ src/main/java/ru/bunkov/calculation/model/calculation/openingproductioncost/PurchasePatent.java
@@ -1,5 +1,6 @@
 package ru.bunkov.calculation.model.calculation.openingproductioncost;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 import lombok.*;
@@ -18,5 +19,5 @@ public class PurchasePatent {
 
     private UUID potentialIncomeId;
 
-    private Double patentCost;
+    private BigDecimal patentCost;
 }
