+++ src/main/java/ru/bunkov/calculation/service/building/argument/CreateBuildingCostArgument.java
@@ -3,11 +3,13 @@ package ru.bunkov.calculation.service.building.argument;
 import lombok.Builder;
 import lombok.Value;
 
+import java.math.BigDecimal;
+
 @Value
 @Builder
 public class CreateBuildingCostArgument {
 
    String type;
 
-   Double cost;
+   BigDecimal cost;
 }
