+++ src/main/java/ru/bunkov/calculation/service/capitalbuilding/argument/SearchCapitalBuildingObjArgument.java
@@ -3,10 +3,12 @@ package ru.bunkov.calculation.service.capitalbuilding.argument;
 import lombok.Builder;
 import lombok.Value;
 
+import java.math.BigDecimal;
+
 @Value
 @Builder
 public class SearchCapitalBuildingObjArgument {
     String type;
 
-    Double cost;
+    BigDecimal cost;
 }
