+++ src/main/java/ru/bunkov/calculation/service/equipment/argument/CreateEquipmentArgument.java
@@ -3,10 +3,12 @@ package ru.bunkov.calculation.service.equipment.argument;
 import lombok.Builder;
 import lombok.Value;
 
+import java.math.BigDecimal;
+
 @Value
 @Builder
 public class CreateEquipmentArgument {
     String equipmentType;
 
-    Double averageCost;
+    BigDecimal averageCost;
 }
