+++ src/main/java/ru/bunkov/calculation/service/income/argument/CreatePotentialIncomeArgument.java
@@ -3,6 +3,7 @@ package ru.bunkov.calculation.service.income.argument;
 import lombok.Builder;
 import lombok.Value;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Value
@@ -11,10 +12,10 @@ public class CreatePotentialIncomeArgument {
 
     String typeOfBusiness;
 
-    Double potentialAnnualIncome;
+    BigDecimal potentialAnnualIncome;
 
-    Double rate;
+    BigDecimal rate;
 
-    Double totalPotentCost;
+    BigDecimal totalPotentCost;
 
 }
