+++ src/main/java/ru/bunkov/calculation/service/income/argument/SearchPotentialIncomeArgument.java
@@ -3,15 +3,16 @@ package ru.bunkov.calculation.service.income.argument;
 import lombok.Builder;
 import lombok.Value;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Value
 @Builder
 public class SearchPotentialIncomeArgument {
     String typeOfBusiness;
-    Double potentialAnnualIncome;
+    BigDecimal potentialAnnualIncome;
 
-    Double rate;
+    BigDecimal rate;
 
-    Double totalPotentCost;
+    BigDecimal totalPotentCost;
 }
