+++ src/main/java/ru/bunkov/calculation/service/region/argument/SearchRegionArgument.java
@@ -3,17 +3,19 @@ package ru.bunkov.calculation.service.region.argument;
 import lombok.Builder;
 import lombok.Value;
 
+import java.math.BigDecimal;
+
 @Value
 @Builder
 public class SearchRegionArgument {
 
     String name;
 
-    Double purchaseLand;
+    BigDecimal purchaseLand;
 
-    Double landLease;
+    BigDecimal landLease;
 
-    Double rentRoomMin;
+    BigDecimal rentRoomMin;
 
-    Double rentRoomMax;
+    BigDecimal rentRoomMax;
 }
